Added config folder to handle growing # of configs
This was inspired by the addition of aws_config.py in an in-progress
branch. That config is needed in multiple directories. To keep things
neat I added a folder under ceruleanserver to hold all (or most)
configs. This will necessitate that
all code that uses these configs is either
a) in ceruleanserver; to this end I will move the model_training code
(in-progress branch) into ceruleanserver
or b) adds ceruleanserver to the path; our pytest code already does this.
NB: move your gitignored configs (local_config.py,
future aws_config.py) into here as well.